Rework - Fixed If-Modified-Since comparison to use second precision for more precision - Removed the GetActionTest class and moved the test to BasicOpsTest class - Updated the test to reuse the server-provided Last-Modified header instead of generating a local timestamp and removed the timing-based sleep.
